Recognise individual items in List resource
It's not enough to be able to update images in a List resource, we
must be able to recognise them in the first place.

The Go YAML parser makes this a little tricky, because it doesn't give
you access to the bytes (like `UnmarshalJSON` does), so if we want
typed resources, we have to parse them generically and marshal them
back out to bytes.