Simplify kubeyaml use with multistage Dockerfile
Instead of using an initContainer in the deployment to copy files
over, we can simply build the files into the image using a multistage
Dockerfile.

This removes the need for a `PLUGINS_PATH` env entry, since we know
where the binaries are at the time the image is built, and can put
that on `$PATH`.