Deduplicate parsing used when applying annotations
To apply the special "tag all" annotation, we need the names of the
containers in a controller. Instead of having a struct definition to
parse just for this, use the existing resource parsing, extending it
just enough to get the containers from things that have containers.

The prior code also failed to take into account List resources and
multidocs; this is no longer a problem since the resource package
already accounts for them.